refactor(deps): move 4 helpers from requirementslib.py to dependencies.py
T_E.3: Per the T_E.1 design + sign-off, move four single-line/short
helpers from pipenv/utils/requirementslib.py to pipenv/utils/dependencies.py
(the canonical home). merge_items brings its two private helpers
(_merge_into, _new_container_like) with it.

Moved:
  is_vcs                     (~14 lines)
  add_ssh_scheme_to_git_uri  (~14 lines)
  merge_items                (~33 lines, + 2 private helpers)
  get_pip_command            (~9 lines)

Caller imports migrated in the same PR per T_C.3 sec.9 / T_E.1
sign-off -- no backwards-compat shim. Five caller files updated:
pipenv/utils/{locking,pipfile}.py (folded into the existing
dependencies-import block), pipenv/project.py (one-line late-import
in _get_vcs_packages, narrow-scoped), tests/unit/test_requirementslib.py
(docstring + import re-target), tests/unit/test_utils.py (the
test_is_vcs late import).

requirementslib.py shrank from 275 to 144 lines (-131). After T_E.3
only one in-tree importer of requirementslib remains
(dependencies.py for unpack_url); T_E.4 will close that out by
relocating the unpack_url/get_http_url fork pair to a new
pipenv/utils/unpack.py and deleting requirementslib.py.

Validation:
- tests/unit/test_dependencies_bridges.py: 31 tests pass (21 from
  T_E.2 + 10 new for T_E.3: 4 import-shape pins, 1 source-cleared
  sanity check, 5 light behaviour pins).
- All four moved symbols importable from pipenv.utils.dependencies;
  no longer exist on pipenv.utils.requirementslib.
- The pipenv/project.py one-line import-line edit is intentionally
  not staged in this commit; T_D.4 is concurrently rewriting that
  file and will sweep up the import-line change as part of its
  larger commit (same pattern as T_D.3 sweeping up T_E.2's
  routines/install.py changes).
