+++ tests/unit/test_utils.py
@@ -282,7 +282,7 @@ class TestUtils:
     )
     @pytest.mark.vcs
     def test_is_vcs(self, entry, expected):
-        from pipenv.utils.requirementslib import is_vcs
+        from pipenv.utils.dependencies import is_vcs
 
         assert is_vcs(entry) is expected
